LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.rect.nonfinite-expected.txt

2d.path.rect.nonfinite

rect() with Infinity/NaN is ignored


PASS rect() with Infinity/NaN is ignored 
